requirements.txt
streamlit>=1.28.0
boto3>=1.28.0
pandas>=2.0.0
openai>=1.0.0
google-genai>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
elevenlabs>=1.0.0
requests>=2.28.0
